+++ ingest/importer/spreadsheet/ingest_workbook.py
@@ -1,7 +1,7 @@
 from openpyxl import Workbook, load_workbook
 
 from ingest.importer.spreadsheet.ingest_worksheet import IngestWorksheet
-from ingest.importer.submission import Submission
+from ingest.importer.submission.submission import Submission
 
 SCHEMAS_WORKSHEET = 'Schemas'
 SPECIAL_TABS = [SCHEMAS_WORKSHEET]
