+++ tests/unit/importer/spreadsheet/test_ingest_workbook.py
@@ -4,7 +4,7 @@ from mock import Mock
 from openpyxl import Workbook
 
 from ingest.importer.spreadsheet.ingest_workbook import IngestWorkbook
-from ingest.importer.submission import Entity
+from ingest.importer.submission.entity import Entity
 from tests.utils import create_ingest_workbook, create_test_workbook, \
     create_mocked_workbook_importer
